requirements.txt
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
fastapi-mail==1.4.1
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
